db_populator/requirements.txt
fandom-py==0.2.1
psycopg2-binary>=2.9.9
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
lxml>=5.3.0 